feat: shared language matching and subtitle track selection
